Indexing issues with xkappa and xp
Prior setup for the disjunctions (two disjunctions and logical constraints connecting their disjuncts) could not work due to indexing differences between xgamma and xkappa/xp. This attempt here (adding V1_ index on xkappa and xp) makes the problem much larger, was giving a different solution and not worth the effort of debugging. Hence, memorizing it to focus on the single disjunction setup